split apart ws connection and auth
