+++ src/lib/test/signalk.rs
@@ -14,7 +14,7 @@ pub struct Config {
     #[default("")]
     wifi_psk: &'static str,
     #[default("")]
-    server_root: &'static str
+    server_root: &'static str,
 }
 
 fn main() -> Result<()> {
@@ -29,5 +29,9 @@ fn main() -> Result<()> {
 
     let app_config = CONFIG;
 
-    signalk_server(app_config.wifi_ssid, app_config.wifi_psk, app_config.server_root)
+    signalk_server(
+        app_config.wifi_ssid,
+        app_config.wifi_psk,
+        app_config.server_root,
+    )
 }
